Fix defining object finalizer in Ruby 3.1 to reference a UUID string
In Ruby 3.1, accessing callback_queue from a finalizer causes an "undefined
local variable or method `callback_queue' for FiniteMachine::Observer:Class
(NameError)", and callback_queue is never cleaned up. However, making the
finalizer an instance method rather than a class method throws a warning
about the improper use of finalizers. Hence, using the UUID string as
an object to check before it is destroyed by garbage collection for when
to call the finalizer proc.

---------

Co-authored-by: Piotr Murach <pmurach@gmail.com>